+++ src/js/core/svg.js
@@ -145,9 +145,7 @@ export default function (UIkit) {
             }
 
             if (this.svg) {
-                this.svg.then(svg => {
-                    svg && svg.remove();
-                });
+                this.svg.then(svg => svg && svg.remove());
                 this.svg = null;
             }
         },